new BeardTrim Project
So my progress so far is that I have created an algorithm for the places the vertices of the face mesh and so far manually labled the beard line indices.

Remember for next time check bookmarks for face mesh indices

also remember we want to conntect the indices to fill an area and then from there we can do beard fade tutorial.